backend/requirements.txt
langchain
langchain-ollama
langchain-community
flask
flask-cors
watchdog   
langchain-openai
python-dotenv